+++ packages/at_client/lib/src/manager/at_client_manager.dart
@@ -1,3 +1,4 @@
+import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/listener/at_sign_change_listener.dart';
 import 'package:at_client/src/listener/switch_at_sign_event.dart';
@@ -52,8 +53,12 @@ class AtClientManager {
     }
   }
 
+  /// * Can provide an [AtServiceFactory] to control what types of AtClients, NotificationServices and SyncServices get created
+  /// by this method.
+  /// * Can provide an [AtChops] instance if available; setCurrentAtSign will ensure that it is injected into objects that
+  /// can use it
   Future<AtClientManager> setCurrentAtSign(String atSign, String? namespace, AtClientPreference preference,
-      {AtServiceFactory? serviceFactory}) async {
+      {AtServiceFactory? serviceFactory, AtChops? atChops}) async {
 
     serviceFactory ??= DefaultAtServiceFactory();
 
@@ -73,6 +78,7 @@ class AtClientManager {
     _atSign = atSign;
     var previousAtClient = _currentAtClient;
     _currentAtClient = await serviceFactory.atClient(_atSign, namespace, preference, this);
+    _currentAtClient!.atChops = atChops;
 
     final switchAtSignEvent =
         SwitchAtSignEvent(previousAtClient, _currentAtClient!);