+++ packages/at_client/CHANGELOG.md
@@ -1,7 +1,6 @@
 ## 3.0.52
 - feat: Introduce AtServiceFactory to make AtClientManager more reusable and more testable
-- feat: Introduce commit log compaction to keep size of commit log thin
-- fix: Fixed a bug where switch atSign event is notified multiple times
+- feat: Make AtChops instance (if any) available everywhere that it can/should be used
 ## 3.0.51
 - feat: Add atSign to AtSignLoggers' names when relevant, so that log messages are clearer
 - feat: Made notificationService and syncService available via AtClient to enable cleaner simpler code elsewhere